# src/backend/MiniCc.Api/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MiniCc.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:5001/"
    }
  ],
  "paths": {
    "/api/Articles": {
      "get": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "default": null
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Article"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Articles"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveArticleRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveArticleRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SaveArticleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              }
            }
          }
        }
      }
    },
    "/api/Articles/{id}": {
      "get": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Article2"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Article2"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Article2"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Articles/content": {
      "post": {
        "tags": [
          "Articles"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentFetchResult"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentFetchResult"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ContentFetchResult"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              }
            }
          }
        }
      }
    },
    "/api/Articles/{id}/favorite": {
      "post": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              }
            }
          }
        }
      }
    },
    "/api/Articles/{id}/archive": {
      "post": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Article2"
                }
              }
            }
          }
        }
      }
    },
    "/api/Articles/{id}/highlights": {
      "post": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Highlight"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Highlight"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Highlight"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Highlight"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Highlight"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Highlight"
                }
              }
            }
          }
        }
      }
    },
    "/api/Articles/highlights/{highlightId}": {
      "delete": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "highlightId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Articles/{id}/tags": {
      "post": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTagRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTagRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Tag"
                }
              }
            }
          }
        }
      }
    },
    "/api/Articles/{id}/tags/{tagId}": {
      "delete": {
        "tags": [
          "Articles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "tagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddTagRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Article": {
        "required": [
          "url",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "originContent": {
            "type": "string"
          },
          "readableContent": {
            "type": "string"
          },
          "textContentLegth": {
            "type": "integer",
            "format": "int32"
          },
          "summary": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isArchived": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "imageUrl": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "highlights": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Highlight"
            }
          }
        }
      },
      "Article2": {
        "required": [
          "url",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "originContent": {
            "type": "string"
          },
          "readableContent": {
            "type": "string"
          },
          "textContentLegth": {
            "type": "integer",
            "format": "int32"
          },
          "summary": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isArchived": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "imageUrl": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag2"
            }
          },
          "highlights": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Highlight"
            }
          }
        }
      },
      "ContentFetchResult": {
        "required": [
          "url",
          "title",
          "originContent"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "originContent": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          }
        }
      },
      "Highlight": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "startOffset": {
            "type": "integer",
            "format": "int32"
          },
          "endOffset": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "articleId": {
            "type": "integer",
            "format": "int32"
          },
          "article": {
            "$ref": "#/components/schemas/Article"
          }
        }
      },
      "SaveArticleRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        }
      },
      "Tag": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "articles": {
            "type": "array",
            "items": { }
          }
        }
      },
      "Tag2": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "articles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Article"
            }
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Articles"
    }
  ]
}